+++ apps/common/frontend/src/components/Dso.tsx
@@ -259,7 +259,7 @@ const DsoViewPrettyJSON: React.FC<DsoViewPrettyJSONProps> = ({
           <Tab label="DSO Info" {...tabProps('dso-info')} />
           <Tab
             label={`${window.splice_config.spliceInstanceNames?.amuletName} Info`}
-            {...tabProps('cc-info')}
+            {...tabProps('amulet-info')}
           />
           {cometBftDebugTab && <Tab label="CometBFT Debug Info" {...tabProps('cometBft-debug')} />}
           {sequencerStatusQuery && mediatorStatusQuery && (